--- a/doc/source/reference/routines.err.rst
+++ b/doc/source/reference/routines.err.rst
@@ -1,3 +1,5 @@
+.. _routines.err:
+
 Floating point error handling
 *****************************