--- a/numpy/core/src/umath/simd.inc.src
+++ b/numpy/core/src/umath/simd.inc.src
@@ -493,12 +493,10 @@ sse2_@kind@_@TYPE@(npy_bool * op, @type@ * ip1, npy_intp n)
 #if @var@ != 0 /* isinf/isfinite */
     /* signbit mask 0x7FFFFFFF after andnot */
     const @vtype@ mask = @vpre@_set1_@vsuf@(-0.@c@);
-    const @vtype@ ones = @vpre@_cmpeq_@vsuf@(@vpre@_setzero_@vsuf@(),
-                                             @vpre@_setzero_@vsuf@());
 #if @double@
-    const @vtype@ fltmax = @vpre@_set1_@vsuf@(DBL_MAX);
+    const @vtype@ infinity = @vpre@_set1_@vsuf@(NPY_INFINITY);
 #else
-    const @vtype@ fltmax = @vpre@_set1_@vsuf@(FLT_MAX);
+    const @vtype@ infinity = @vpre@_set1_@vsuf@(NPY_INFINITY);
 #endif
 #endif
     LOOP_BLOCK_ALIGN_VAR(ip1, @type@, VECTOR_SIZE_BYTES) {
@@ -510,38 +508,39 @@ sse2_@kind@_@TYPE@(npy_bool * op, @type@ * ip1, npy_intp n)
         @vtype@ c = @vpre@_load_@vsuf@(&ip1[i + 2 * VECTOR_SIZE_BYTES / sizeof(@type@)]);
         @vtype@ d = @vpre@_load_@vsuf@(&ip1[i + 3 * VECTOR_SIZE_BYTES / sizeof(@type@)]);
         @vtype@ r1, r2, r3, r4;
+
 #if @var@ != 0 /* isinf/isfinite */
-        /* fabs via masking of sign bit */
+        /* Convert -infinity to infinity: fabs via masking of sign bit */
         r1 = @vpre@_andnot_@vsuf@(mask, a);
         r2 = @vpre@_andnot_@vsuf@(mask, b);
         r3 = @vpre@_andnot_@vsuf@(mask, c);
         r4 = @vpre@_andnot_@vsuf@(mask, d);
-#if @var@ == 1 /* isfinite */
-        /* negative compare against max float, nan is always true */
-        r1 = @vpre@_cmpnle_@vsuf@(r1, fltmax);
-        r2 = @vpre@_cmpnle_@vsuf@(r2, fltmax);
-        r3 = @vpre@_cmpnle_@vsuf@(r3, fltmax);
-        r4 = @vpre@_cmpnle_@vsuf@(r4, fltmax);
-#else /* isinf */
-        r1 = @vpre@_cmpnlt_@vsuf@(fltmax, r1);
-        r2 = @vpre@_cmpnlt_@vsuf@(fltmax, r2);
-        r3 = @vpre@_cmpnlt_@vsuf@(fltmax, r3);
-        r4 = @vpre@_cmpnlt_@vsuf@(fltmax, r4);
 #endif
-        /* flip results to what we want (andnot as there is no sse not) */
-        r1 = @vpre@_andnot_@vsuf@(r1, ones);
-        r2 = @vpre@_andnot_@vsuf@(r2, ones);
-        r3 = @vpre@_andnot_@vsuf@(r3, ones);
-        r4 = @vpre@_andnot_@vsuf@(r4, ones);
-#endif
+
 #if @var@ == 0 /* isnan */
         r1 = @vpre@_cmpneq_@vsuf@(a, a);
         r2 = @vpre@_cmpneq_@vsuf@(b, b);
         r3 = @vpre@_cmpneq_@vsuf@(c, c);
         r4 = @vpre@_cmpneq_@vsuf@(d, d);
+#elif @var@ == 1 /* isfinite */
+        /* compare against infinity, nan is always false */
+        r1 = @vpre@_cmplt_@vsuf@(r1, infinity);
+        r2 = @vpre@_cmplt_@vsuf@(r2, infinity);
+        r3 = @vpre@_cmplt_@vsuf@(r3, infinity);
+        r4 = @vpre@_cmplt_@vsuf@(r4, infinity);
+#elif @var@ == 2  /* isinf */
+        r1 = @vpre@_cmpeq_@vsuf@(infinity, r1);
+        r2 = @vpre@_cmpeq_@vsuf@(infinity, r2);
+        r3 = @vpre@_cmpeq_@vsuf@(infinity, r3);
+        r4 = @vpre@_cmpeq_@vsuf@(infinity, r4);
 #endif
         sse2_compress4_to_byte_@TYPE@(r1, r2, r3, &r4, &op[i]);
     }
+
+#if @var@ == 1  /* isfinite */
+    /* `isfinite` uses `cmplt`, which will set invalid value FPE for NaNs. */
+    npy_clear_floatstatus_barrier((char *)op);
+#endif
     LOOP_BLOCKED_END {
         op[i] = npy_@kind@(ip1[i]) != 0;
     }
@@ -592,6 +591,7 @@ sse2_binary_@kind@_@TYPE@(npy_bool * op, @type@ * ip1, @type@ * ip2, npy_intp n)
     LOOP_BLOCKED_END {
         op[i] = sse2_ordered_cmp_@kind@_@TYPE@(ip1[i], ip2[i]);
     }
+    npy_clear_floatstatus_barrier((char *)op);
 }
 
 
@@ -616,6 +616,7 @@ sse2_binary_scalar1_@kind@_@TYPE@(npy_bool * op, @type@ * ip1, @type@ * ip2, npy
     LOOP_BLOCKED_END {
         op[i] = sse2_ordered_cmp_@kind@_@TYPE@(ip1[0], ip2[i]);
     }
+    npy_clear_floatstatus_barrier((char *)op);
 }
 
 
@@ -640,6 +641,7 @@ sse2_binary_scalar2_@kind@_@TYPE@(npy_bool * op, @type@ * ip1, @type@ * ip2, npy
     LOOP_BLOCKED_END {
         op[i] = sse2_ordered_cmp_@kind@_@TYPE@(ip1[i], ip2[0]);
     }
+    npy_clear_floatstatus_barrier((char *)op);
 }
 /**end repeat1**/