@seberg 2021-07-14
Well, that attempt died quickly on CI :/, it had worked great locally.  (I still have some hope I just missed some SIMD path though)