@seberg 2021-07-14
Aha, I managed to rewrite the AVX512F logic to avoid the spurious warning setting.  Quick timing gives (on a machine with `AVX512F*`):
```
In [1]: arr1 = np.random.random(10000); arr2 = np.random.random(10000)
```
On main:
```
In [2]: %timeit np.minimum(arr1, arr2)
7.89 µs ± 83.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
On this branch:
```
In [3]: %timeit np.minimum(arr1, arr2)
8.39 µs ± 6.36 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
So it looks like <5% slowdown, probably we can live with that.

---

I *think* the AVX512 solution may be a big improvement for the SSE2 code.  That code currently tests for an invalid flag set during the comparison.  But as far as I can tell, both of the instructions I now use in AVX have equivalent ones in SSE2.