@seberg 2021-07-14
I can reproduce the MacOS failures locally with clang, and they go away with `-ffp-exception-behavior=strict`, so gh-19049 (not working) strikes again...

So we will have to figure out gh-19049 before this has any chance :(.  Marking as draft, I guess some of the SIMD stuff could be split out regardless of clang.