@seberg 2021-07-19
Well, probably will have to just pull out the useful parts...  Godbolt points out that `clang <10.0` does bad optimization: https://godbolt.org/z/38hcaaP7o (note the `cmpltsd xmm3, xmm1` instruction, where `clang 10` correctly only uses `ucomisd`). 

I am not actually sure why this fails CI, since that looks like a clang 11, unless it is Mac specific?  But, I guess we support clang 9 anyway, so...