@seberg 2021-11-10
Closing reopening, because I am curious if bumping MacOS in CI from 10.14 to 10.15 makes a difference here (although I am not sure that would actually help with moving forward).