@seberg 2022-04-11
Going to close this for now.  There are a few things here that would be nice to have.  The whole thing would be nice, but seems only viable for newer clang versions (which probably takes a bit more time).